requirements.txt
python-telegram-bot==21.0
python-dotenv==1.0.1
pydantic==2.6.1
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
supabase==2.3.4
httpx==0.27.0
better-profanity==0.7.0
